Fix Layer paths for local repos
The current implementation of the layer push code replaces the repository endpoint with the local IP. This
works great for images that are being pulled from docker hub or have no repository component to their tags.
It fails when the image is local, but there is a registry endpoint. If the remote registry model worked in all
cases, this change would not be necessary. For the time being, it appears that clairctl doesn't work with ECR
credentials, so the `--local` route is necessary.